Fixed "not iterable" exeception.
But for some reason date format still not valid for datetime
fields, getting:

Warning:
The ticket field 'userfinish' is invalid: Field userfinish does
not seem to look like a date. The correct format is YYYY-MM-DD.
The ticket field 'userstart' is invalid: Field userstart does not
seem to look like a date. The correct format is YYYY-MM-DD.